+++ internal/resources/resource_scope.go
@@ -312,7 +312,7 @@ func (r *ScopeResource) Update(ctx context.Context, req resource.UpdateRequest, 
 	})
 }
 
-func (r *ScopeResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *ScopeResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) { //nolint:dupl
 	var state ScopeConfig
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {