+++ internal/resources/resource_scope_group.go
@@ -214,7 +214,7 @@ func (r *ScopeGroupResource) Update(ctx context.Context, req resource.UpdateRequ
 	})
 }
 
-func (r *ScopeGroupResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *ScopeGroupResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) { //nolint:dupl
 	var state ScopeGroupConfig
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {