+++ internal/resources/resource_template_group.go
@@ -324,7 +324,7 @@ func (r *TemplateGroupResource) Update(ctx context.Context, req resource.UpdateR
 	})
 }
 
-func (r *TemplateGroupResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+func (r *TemplateGroupResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) { //nolint:dupl
 	var state TemplateGroupConfig
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {